:recycle: Add OP_CODE files
