+++ BitcoinKit/Opcode.swift
@@ -20,11 +20,6 @@
 
 import Foundation
 
-// Constants
-private let blobFalse: Data = Data()
-private let blobZero: Data = Data()
-private let blobTrue: Data = Data(bytes: [UInt8(1)])
-
 public class OpCode: OpCodeProtocol {
     public var value: UInt8 { return 0x00 }
     public var name: String { return "" }
@@ -305,7 +300,7 @@ public class OpCode: OpCodeProtocol {
         OP_INVALIDOPCODE
     ]
 
-    fileprivate init() {}
+    internal init() {}
 }
 
 extension OpCode: Comparable {
@@ -318,121 +313,6 @@ extension OpCode: Comparable {
     }
 }
 
-public class ScriptExecutionContext {
-    // Flags affecting verification. Default is the most liberal verification.
-    // One can be stricter to not relay transactions with non-canonical signatures and pubkey (as BitcoinQT does).
-    // Defaults in CoreBitcoin: be liberal in what you accept and conservative in what you send.
-    // So we try to create canonical purist transactions but have no problem accepting and working with non-canonical ones.
-    public var verificationFlags: ScriptVerification?
-
-    // Stack contains Data objects that are interpreted as numbers, bignums, booleans or raw data when needed.
-    public fileprivate(set) var stack = [Data]()
-    // Used in ALTSTACK ops.
-    public fileprivate(set) var altStack = [Data]()
-    // Holds an array of Bool values to keep track of if/else branches.
-    public fileprivate(set) var conditionStack = [Bool]()
-
-    // Currently executed script.
-    public fileprivate(set) var script: Script = Script()
-    // Current opcode.
-    public fileprivate(set) var opCode: OpCode = OpCode.OP_0
-    // Current payload for any "push data" operation.
-    // public var data
-    // Current opcode index in _script.
-    public fileprivate(set) var opIndex: Int = 0
-    // Index of last OP_CODESEPARATOR
-    public fileprivate(set) var lastCodeSepartorIndex: Int = 0
-
-    // Keeps number of executed operations to check for limit.
-    public var opCount: Int = 0
-
-    // Transaction, utxo, index for CHECKSIG operations
-    public var transaction: Transaction?
-    public var utxoToVerify: TransactionOutput?
-    public var inputIndex: UInt32 = 0xffffffff
-
-    public var shouldExecute: Bool {
-        return !conditionStack.contains(false)
-    }
-
-    func normalized(_ index: Int) -> Int {
-        return (index < 0) ? stack.count + index : index
-    }
-
-    internal func pushData(_ data: Data) throws {
-        guard data.count <= BTC_MAX_SCRIPT_ELEMENT_SIZE else {
-            throw ScriptMachineError.error("PushedData size is too big.")
-        }
-        stack.append(data)
-    }
-
-    internal func resetStack() {
-        stack = [Data()]
-        altStack = [Data()]
-        conditionStack = [Bool]()
-    }
-
-    internal func swapDataAt(i: Int, j: Int) {
-        stack.swapAt(normalized(i), normalized(j))
-    }
-
-    internal func deserializeP2SHLockScript() throws -> Script {
-        // Make a copy of the stack if we have P2SH script.
-        // We will run deserialized P2SH script on this stack.
-        var stackForP2SH: [Data] = stack
-
-        // Instantiate the script from the last data on the stack.
-        guard let last = stackForP2SH.last, let deserializedLockScript = Script(data: last) else {
-            // stackForP2SH cannot be empty here, because if it was the
-            // P2SH  HASH <> EQUAL  scriptPubKey would be evaluated with
-            // an empty stack and the runScript: above would return NO.
-            throw ScriptMachineError.exception("internal inconsistency: stackForP2SH cannot be empty at this point.")
-        }
-
-        // Remove it from the stack.
-        stackForP2SH.removeLast()
-
-        // Replace current stack with P2SH stack.
-        resetStack()
-        stack = stackForP2SH
-        return deserializedLockScript
-    }
-
-    public func data(at i: Int) -> Data {
-        return stack[normalized(i)]
-    }
-
-    public func number(at i: Int) -> Int32? {
-        let data: Data = stack[normalized(i)]
-        if data.count > 4 {
-            return nil
-        }
-        return Int32(data.withUnsafeBytes { $0.pointee })
-    }
-
-    public func bool(at i: Int) -> Bool {
-        let data: Data = stack[normalized(i)]
-        guard !data.isEmpty else {
-            return false
-        }
-
-        for (i, byte) in data.enumerated() where byte != 0 {
-            // Can be negative zero, also counts as false
-            if i == (data.count - 1) && byte == 0x80 {
-                return false
-            }
-            return true
-        }
-        return false
-    }
-}
-
-extension ScriptExecutionContext: CustomStringConvertible {
-    public var description: String {
-        return ""
-    }
-}
-
 // TODO: Remove this when all opcodes are implemented
 public class OpExample: OpCode {
     override public var value: UInt8 { return 0x00 }
@@ -442,150 +322,5 @@ public class OpExample: OpCode {
         try super.execute(context)
         // do something with context here!
         fatalError("You should implement this!")
-    }
-}
-public class Op0: OpCode {
-    override public var value: UInt8 { return 0x00 }
-    override public var name: String { return "OP_0" }
-}
-
-public class OpN: OpCode {
-    override public var value: UInt8 { return 0x50 + n }
-    override public var name: String { return "OP_\(n)" }
-    private let n: UInt8
-    fileprivate init(_ n: UInt8) {
-        guard (1...16).contains(n) else {
-            fatalError("OP_N can be initialized with N between 1 and 16. \(n) is not valid.")
-        }
-        self.n = n
-        super.init()
-    }
-}
-
-public class OpPushData1: OpCode {
-    override public var value: UInt8 { return 0x4c }
-    override public var name: String { return "OP_PUSHDATA1" }
-}
-public class OpPushData2: OpCode {
-    override public var value: UInt8 { return 0x4d }
-    override public var name: String { return "OP_PUSHDATA2" }
-}
-public class OpPushData4: OpCode {
-    override public var value: UInt8 { return 0x4e }
-    override public var name: String { return "OP_PUSHDATA4" }
-}
-
-public class OpVerify: OpCode {
-    override public var value: UInt8 { return 0x69 }
-    override public var name: String { return "OP_VERIFY" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        // (true -- ) or
-        // (false -- false) and return
-        guard context.stack.count >= 1 else {
-            throw ScriptMachineError.opcodeRequiresItemsOnStack(1)
-        }
-        guard context.bool(at: -1) else {
-            throw ScriptMachineError.error("OP_VERIFY failed.")
-        }
-        context.stack.removeLast()
-    }
-}
-
-public class OpDuplicate: OpCode {
-    override public var value: UInt8 { return 0x76 }
-    override public var name: String { return "OP_DUP" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        // (x -- x x)
-        guard context.stack.count >= 1 else {
-            throw ScriptMachineError.opcodeRequiresItemsOnStack(1)
-        }
-        context.stack.append(context.data(at: -1))
-    }
-}
-
-public class OpHash160: OpCode {
-    override public var value: UInt8 { return 0xa9 }
-    override public var name: String { return "OP_HASH160" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        // (in -- hash)
-        guard context.stack.count >= 1 else {
-            throw ScriptMachineError.opcodeRequiresItemsOnStack(1)
-        }
-
-        let data: Data = context.stack.removeLast()
-        let hash: Data = Crypto.sha256ripemd160(data)
-        context.stack.append(hash)
-    }
-}
-
-public class OpEqual: OpCode {
-    override public var value: UInt8 { return 0x87 }
-    override public var name: String { return "OP_EQUAL" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        // (x1 x2 - bool)
-        guard context.stack.count >= 2 else {
-            throw ScriptMachineError.opcodeRequiresItemsOnStack(2)
-        }
-
-        let x1 = context.stack.popLast()!
-        let x2 = context.stack.popLast()!
-        let equal: Bool = x1 == x2
-        context.stack.append(equal ? blobTrue : blobFalse)
-    }
-}
-
-public class OpEqualVerify: OpCode {
-    override public var value: UInt8 { return 0x88 }
-    override public var name: String { return "OP_EQUALVERIFY" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        print("stack: \(context.stack.map { $0.hex }.joined(separator: " "))")
-        try OpCode.OP_EQUAL.execute(context)
-        try OpCode.OP_VERIFY.execute(context)
-    }
-}
-
-public class OpCheckSig: OpCode {
-    override public var value: UInt8 { return 0xac }
-    override public var name: String { return "OP_CHECKSIGVERIFY" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        guard context.stack.count >= 2 else {
-            throw ScriptMachineError.opcodeRequiresItemsOnStack(2)
-        }
-        print("stack: \(context.stack.map { $0.hex }.joined(separator: " "))")
-
-        let pubkeyData: Data = context.stack.removeLast()
-        let sigData: Data = context.stack.removeLast()
-
-        // Subset of script starting at the most recent OP_CODESEPARATOR (inclusive)
-        let subScript = context.script.subScript(from: context.lastCodeSepartorIndex)
-        subScript.deleteOccurrences(of: sigData)
-
-        guard let tx = context.transaction, let utxo = context.utxoToVerify else {
-            throw ScriptMachineError.error("The transaction or the utxo to verify is not set.")
-        }
-        let valid = try Crypto.verifySigData(for: tx, inputIndex: Int(context.inputIndex), utxo: utxo, sigData: sigData, pubKeyData: pubkeyData)
-        context.stack.append(valid ? blobTrue : blobFalse)
-    }
-}
-
-public class OpInvalidOpCode: OpCode {
-    override public var value: UInt8 { return 0xff }
-    override public var name: String { return "OP_INVALIDOPCODE" }
-
-    override public func execute(_ context: ScriptExecutionContext) throws {
-        try super.execute(context)
-        // do something with context here!
     }
 }
