+++ BitcoinKit/ScriptMachine.swift
@@ -40,17 +40,6 @@ public enum ScriptMachineError: Error {
 // returning a bool indicating valid or not. There are no loops.
 // You can -copy a machine which will copy all the parameters and the stack state.
 class ScriptMachine {
-
-    // Constants
-    private let blobFalse: Data = Data()
-    private let blobZero: Data = Data()
-    private let blobTrue: Data = Data(bytes: [UInt8(1)])
-    // TODO: check later
-    private let bigNumberZero: Data = Data() + UInt64(0)
-    private let bigNumberOne: Data = Data() + UInt64(1)
-    private let bigNumberFalse: Data = Data() + UInt64(0)
-    private let bigNumberTrue: Data = Data() + UInt64(1)
-
     // "To" transaction that is signed by an inputScript.
     // Required parameter.
     public var transaction: Transaction?
